Proga/homework08-elevennote/requirements/base.txt
Django==3.0.6
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
uWSGI==2.0.18
django-widget-tweaks==1.4.8
django-wysiwyg==0.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-registration